refactor(settings): rename change-card-modal to change-cc-modal
